Use jinja2's trim_blocks and lstrip_blocks
This makes the taking care of white-spaces much more intuitive,
because "minus" signs in blocks '{%- -%}' are not needed
(even though sometimes those are still needed in '{{ -}}').  The
language is still complete but see how the 'docker.tpl' template
was simplified in this commit.

To revert to the previous behavior users can use
--keep-block-whitespaces option.