If unallowed combination of selectors, exit with 2
We need that for example for container images where we run
distgen for all combinations of distros/versions and some of
those combinations might not be allowed or present in multispec.
In this specific case, we want to silence the output from distgen
so the log from the generator is not full of errors.