Fix builds on EPELs due to outdated components
Some of the new functionalities that are present in Fedora and are
needed for distgen to build using the refreshed build system are not
present in EPEL.

This means that we have to rollback some changes to allow build on
EPEL. This shouldn't impact functionality of the application, but
keeping it in the main repository would create mess in the codebase.
For that reason the rollbacks are handled by patches in spec file.